complete_setup
