+++ cmd/eval_conf
@@ -1,0 +1,32 @@
+#!/usr/bin/env python3
+
+
+def import_proto_kernel(proto_kernel_rel_path):
+    """
+    `protoprimer` entry script boilerplate function to import `proto_kernel`.
+    """
+    import os
+    import importlib.util
+
+    module_spec = importlib.util.spec_from_file_location(
+        "proto_kernel",
+        os.path.join(
+            os.path.dirname(__file__),
+            proto_kernel_rel_path,
+        ),
+    )
+    loaded_proto_kernel = importlib.util.module_from_spec(module_spec)
+    module_spec.loader.exec_module(loaded_proto_kernel)
+    return loaded_proto_kernel
+
+
+if __name__ == "__main__":
+    proto_kernel = import_proto_kernel("./proto_code/proto_kernel.py")
+
+    import os
+
+    os.environ["PROTOPRIMER_DO_INSTALL"] = str(False)
+
+    # TODO: TODO_28_48_19_20.api_to_traverse_config_when_primed.md:
+    #       convert from `boot_env` to `start_app` when it can access the config.
+    proto_kernel.boot_env("neoprimer.cmd_eval_conf:custom_main")
