+++ cmd/proto_code/proto_kernel.py
@@ -1630,1574 +1630,8 @@ def str_to_bool(v: str) -> bool:
 
 
 ########################################################################################################################
-# Visitors for config nodes.
-# See: FT_19_44_42_19.effective_config.md
 
 
-class AbstractConfigVisitor:
-    """
-    Implements the visitor pattern for classed derived from `AbstractConfigNode`.
-    """
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def visit_dict(
-        self,
-        dict_node: "AbstractDictNode",
-        **kwargs,
-    ) -> None:
-        pass
-
-    def visit_list(
-        self,
-        list_node: "AbstractListNode",
-        **kwargs,
-    ) -> None:
-        pass
-
-    def visit_value(
-        self,
-        value_node: "AbstractValueNode",
-        **kwargs,
-    ) -> None:
-        pass
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def visit_root(
-        self,
-        root_node: "AbstractRootNode",
-        **kwargs,
-    ) -> None:
-        pass
-
-
-class RenderConfigVisitor(AbstractConfigVisitor):
-    """
-    Render a JSON-like data structure into data coded in `python` (with annotations as comments).
-
-    It renders loaded config as: FT_19_44_42_19.effective_config.md
-    """
-
-    def __init__(
-        self,
-        is_quiet: bool = False,
-    ):
-        self.is_quiet: bool = is_quiet
-        self.rendered_value: str = ""
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def render_node(
-        self,
-        config_node: "AbstractConfigNode",
-    ) -> str:
-        s: str = ""
-        if not self.is_quiet:
-            s += " " * config_node.node_indent + os.linesep
-        s += self._render_node_annotation(config_node)
-
-        rendered_value: str = self._render_node_value(config_node)
-        if config_node.is_present:
-            s += rendered_value
-        else:
-            s += self._comment_with_indent(
-                rendered_value,
-                config_node,
-            )
-        return s
-
-    def _render_node_annotation(
-        self,
-        config_node: "AbstractConfigNode",
-    ) -> str:
-        if self.is_quiet:
-            return ""
-        note_text = config_node.note_text
-        if len(note_text.strip()) == 0:
-            return ""
-        annotation_lines = note_text.splitlines()
-        s = ""
-        for annotation_line in annotation_lines:
-            s += (
-                " " * config_node.node_indent
-                + f"{config_node.note_color.value}# {annotation_line}{TermColor.reset_style.value}"
-                + os.linesep
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-            )
-        return s
-
-    @staticmethod
-    def _render_node_name(
-        config_node: "AbstractConfigNode",
-    ) -> str:
-        if config_node.node_name is None:
-            return ""
-        return f"{json.dumps(config_node.node_name)}: "
-
-    def _render_node_value(
-        self,
-        config_node: "AbstractConfigNode",
-    ) -> str:
-        config_node.accept_visitor(self)
-        return self.rendered_value
-
-    def _comment_with_indent(
-        self,
-        rendered_text: str,
-        config_node: "AbstractConfigNode",
-    ) -> str:
-        if self.is_quiet:
-            return ""
-        deactivated_lines = []
-        rendered_lines = rendered_text.splitlines()
-        for rendered_line in rendered_lines:
-            deactivated_lines.append(
-                rendered_line[: config_node.node_indent]
-                + f"{config_node.note_color.value}# "
-                + rendered_line[config_node.node_indent :]
-                + f"{TermColor.reset_style.value}"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-            )
-        return os.linesep.join(deactivated_lines)
-
-    def visit_dict(
-        self,
-        dict_node: "AbstractDictNode",
-        **kwargs,
-    ):
-        s: str = ""
-        s += (
-            " " * dict_node.node_indent
-            + self._render_node_name(dict_node)
-            + "{"
-            + os.linesep
-            #
-        )
-        for child_name, child_node in dict_node.child_nodes.items():
-            rendered_child = self.render_node(child_node)
-            s += rendered_child
-            if rendered_child or not self.is_quiet:
-                s += os.linesep
-        s += " " * dict_node.node_indent + "},"
-        self.rendered_value = s
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def visit_list(
-        self,
-        list_node: "AbstractListNode",
-        **kwargs,
-    ):
-        s: str = ""
-        s += (
-            " " * list_node.node_indent
-            + self._render_node_name(list_node)
-            + "["
-            + os.linesep
-            #
-        )
-        for child_node in list_node.child_nodes:
-            rendered_child = self.render_node(child_node)
-            s += rendered_child
-            if rendered_child or not self.is_quiet:
-                s += os.linesep
-        s += " " * list_node.node_indent + "],"
-        self.rendered_value = s
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def visit_value(
-        self,
-        value_node: "AbstractValueNode",
-        **kwargs,
-    ):
-        s: str = ""
-        if isinstance(value_node.orig_data, str):
-            s += (
-                " " * value_node.node_indent
-                + self._render_node_name(value_node)
-                # Use double-quote for `str`:
-                + f"{json.dumps(value_node.orig_data)}"
-                + ","
-                #
-            )
-        else:
-            s += (
-                " " * value_node.node_indent
-                + self._render_node_name(value_node)
-                + f"{repr(value_node.orig_data)}"
-                + ","
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-            )
-        self.rendered_value = s
-
-    def visit_root(
-        self,
-        root_node: "AbstractRootNode",
-        **kwargs,
-    ):
-        # Remove the last char (which is supposed to be `,`):
-        rendered_child: str = self.render_node(root_node.child_node)[:-1]
-        s: str = ""
-        s += " " * root_node.node_indent + f"{root_node.node_name} = (" + os.linesep
-        s += rendered_child + os.linesep
-        s += " " * root_node.node_indent + ")"
-        self.rendered_value = s
-
-
-class ConfigBuilderVisitor(AbstractConfigVisitor):
-    """
-    Builds a config node and visits it to build children config nodes.
-    """
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def build_config_node(
-        self,
-        orig_data: Any,
-        **kwargs,
-    ) -> "AbstractConfigNode":
-        if isinstance(orig_data, dict):
-            return self.build_dict_node(
-                orig_data=orig_data,
-                **kwargs,
-            )
-        elif isinstance(orig_data, list):
-            return self.build_list_node(
-                orig_data=orig_data,
-                **kwargs,
-            )
-        else:
-            return self.build_value_node(
-                orig_data=orig_data,
-                **kwargs,
-            )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def build_dict_node(
-        self,
-        **kwargs,
-    ) -> "AbstractDictNode":
-        kwargs.pop("is_present", None)
-        return AbstractDictNode(
-            is_present=True,
-            child_builder=self,
-            **kwargs,
-        )
-
-    def build_list_node(
-        self,
-        **kwargs,
-    ) -> "AbstractListNode":
-        kwargs.pop("is_present", None)
-        return AbstractListNode(
-            is_present=True,
-            child_builder=self,
-            **kwargs,
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def build_value_node(
-        self,
-        **kwargs,
-    ) -> "AbstractValueNode":
-        kwargs.pop("is_present", None)
-        return AbstractValueNode(
-            is_present=True,
-            **kwargs,
-        )
-
-    def build_root_node(
-        self,
-        **kwargs,
-    ) -> "AbstractRootNode":
-        kwargs.pop("is_present", None)
-        return AbstractRootNode(
-            is_present=True,
-            child_builder=self,
-            **kwargs,
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def visit_dict(
-        self,
-        dict_node: "AbstractDictNode",
-        **kwargs,
-    ) -> None:
-        if dict_node.orig_data is None:
-            return
-        kwargs.pop("node_name", None)
-        kwargs.pop("node_indent", None)
-        kwargs.pop("orig_data", None)
-        for field_name, field_value in dict_node.orig_data.items():
-            child_node = self.build_config_node(
-                node_name=field_name,
-                node_indent=dict_node.node_indent + AbstractConfigNode.indent_size,
-                orig_data=field_value,
-                **kwargs,
-            )
-            dict_node.child_nodes[field_name] = child_node
-
-    def visit_list(
-        self,
-        list_node: "AbstractListNode",
-        **kwargs,
-    ) -> None:
-        if list_node.orig_data is None:
-            return
-        kwargs.pop("node_name", None)
-        kwargs.pop("node_indent", None)
-        kwargs.pop("orig_data", None)
-        for list_item in list_node.orig_data:
-            child_node = self.build_config_node(
-                node_name=None,
-                node_indent=list_node.node_indent + AbstractConfigNode.indent_size,
-                orig_data=list_item,
-                **kwargs,
-            )
-            list_node.child_nodes.append(child_node)
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def visit_value(
-        self,
-        value_node: "AbstractValueNode",
-        **kwargs,
-    ) -> None:
-        # Value nodes have no children.
-        pass
-
-    def visit_root(
-        self,
-        root_node: "AbstractRootNode",
-        **kwargs,
-    ) -> None:
-        if root_node.orig_data is None:
-            return
-        kwargs.pop("node_name", None)
-        kwargs.pop("node_indent", None)
-        kwargs.pop("orig_data", None)
-        root_node.child_node = self.build_config_node(
-            node_name=None,
-            node_indent=root_node.node_indent + AbstractConfigNode.indent_size,
-            orig_data=root_node.orig_data,
-            **kwargs,
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-class AnnotateUnusedVisitor(AbstractConfigVisitor):
-    """
-    Annotates a config node as unused (not recursively).
-    """
-
-    def visit_dict(
-        self,
-        dict_node: "AbstractDictNode",
-        **kwargs,
-    ) -> None:
-        dict_node.note_text = f"This `dict` is not used by the `{ConfConstGeneral.name_protoprimer_package}`."
-        dict_node.note_color = TermColor.config_unused
-
-    def visit_list(
-        self,
-        list_node: "AbstractListNode",
-        **kwargs,
-    ) -> None:
-        list_node.note_text = f"This `list` is not used by the `{ConfConstGeneral.name_protoprimer_package}`."
-        list_node.note_color = TermColor.config_unused
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def visit_value(
-        self,
-        value_node: "AbstractValueNode",
-        **kwargs,
-    ) -> None:
-        value_node.note_text = f"This value is not used by the `{ConfConstGeneral.name_protoprimer_package}`."
-        value_node.note_color = TermColor.config_unused
-
-    def visit_root(
-        self,
-        root_node: "AbstractRootNode",
-        **kwargs,
-    ) -> None:
-        root_node.note_text = f"This config is not used by the `{ConfConstGeneral.name_protoprimer_package}`."
-        root_node.note_color = TermColor.config_unused
-
-
-class UnusedConfigBuilderVisitor(ConfigBuilderVisitor):
-    """
-    Builds a config node (recursively) and annotates that top-level node as unused.
-    """
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def __init__(
-        self,
-    ):
-        self.recursion_level: int = 0
-
-    def build_config_node(
-        self,
-        orig_data: Any,
-        **kwargs,
-    ) -> "AbstractConfigNode":
-        self.recursion_level += 1
-        config_node: AbstractConfigNode = super().build_config_node(
-            orig_data=orig_data,
-            **kwargs,
-        )
-        self.recursion_level -= 1
-        if self.recursion_level == 0:
-            # Set annotation only for the top-level node:
-            config_node.accept_visitor(AnnotateUnusedVisitor())
-        return config_node
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-########################################################################################################################
-# Abstract config node types.
-# See: FT_19_44_42_19.effective_config.md
-
-
-class AbstractConfigNode(Generic[ValueType]):
-    """
-    Models a node in a JSON-like (nested) data structure.
-
-    It loads config from: FT_48_62_07_98.config_format.md
-    """
-
-    indent_size: int = 4
-
-    def __init__(
-        self,
-        node_name: str | None,
-        node_indent: int,
-        is_present: bool,
-        orig_data: ValueType | None,
-        **kwargs,
-    ):
-        self.node_name: str | None = node_name
-        self.node_indent: int = node_indent
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        # Unlike simply setting `orig_data` to `None`, setting `is_present` to `False`
-        # allows distinguishing between (A) a valid `None` value and (B) a missing value.
-        self.is_present: bool = is_present
-
-        self.orig_data: ValueType | None = orig_data
-
-        self.note_text: str = ""
-        self.note_color: TermColor = TermColor.config_comment if is_present else TermColor.config_missing
-
-    def accept_visitor(
-        self,
-        visitor: AbstractConfigVisitor,
-        **kwargs,
-    ) -> None:
-        """
-        Accept a `AbstractConfigVisitor`.
-        """
-        raise NotImplementedError()
-
-
-class AbstractDictNode(AbstractConfigNode[dict]):
-    """
-    Models `{ ... }` JSON-like `dict`.
-    """
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def __init__(
-        self,
-        child_builder: ConfigBuilderVisitor,
-        **kwargs,
-    ):
-        super().__init__(
-            **kwargs,
-        )
-        self.child_nodes: dict[str, AbstractConfigNode] = {}
-        self.accept_visitor(
-            child_builder,
-            **kwargs,
-        )
-
-    def accept_visitor(
-        self,
-        visitor: AbstractConfigVisitor,
-        **kwargs,
-    ) -> None:
-        visitor.visit_dict(
-            self,
-            **kwargs,
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-class AbstractListNode(AbstractConfigNode[list]):
-    """
-    Models `[ ... ]` JSON-like `list`.
-    """
-
-    def __init__(
-        self,
-        child_builder: ConfigBuilderVisitor,
-        **kwargs,
-    ):
-        super().__init__(
-            **kwargs,
-        )
-        self.child_nodes: list[AbstractConfigNode] = []
-        self.accept_visitor(
-            child_builder,
-            **kwargs,
-        )
-
-    def accept_visitor(
-        self,
-        visitor: AbstractConfigVisitor,
-        **kwargs,
-    ) -> None:
-        visitor.visit_list(
-            self,
-            **kwargs,
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-class AbstractValueNode(AbstractConfigNode[ValueType]):
-    """
-    Models any simple value in JSON-like data structure (neither `list` nor `dict`).
-    """
-
-    def __init__(
-        self,
-        orig_data: ValueType | None,
-        **kwargs,
-    ):
-        super().__init__(
-            orig_data=orig_data,
-            **kwargs,
-        )
-        assert not isinstance(orig_data, list)
-        assert not isinstance(orig_data, dict)
-
-    def accept_visitor(
-        self,
-        visitor: AbstractConfigVisitor,
-        **kwargs,
-    ) -> None:
-        visitor.visit_value(
-            self,
-            **kwargs,
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-class AbstractRootNode(AbstractConfigNode[ValueType]):
-    """
-    Wraps any given `child_node` as `node_name = ( child_node )`.
-
-    Syntactically, it represents an assignment to the `node_name` var and
-    allows accessing the assigned data via `compile_effective_config`.
-    """
-
-    def __init__(
-        self,
-        child_builder: ConfigBuilderVisitor,
-        **kwargs,
-    ):
-        super().__init__(
-            **kwargs,
-        )
-        self.child_node: AbstractConfigNode | None = None
-        self.accept_visitor(
-            child_builder,
-            **kwargs,
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def accept_visitor(
-        self,
-        visitor: AbstractConfigVisitor,
-        **kwargs,
-    ) -> None:
-        visitor.visit_root(
-            self,
-            **kwargs,
-        )
-
-    def compile_effective_config(
-        self,
-    ) -> Any:
-        """
-        Produces rendered config and compiles it to access data.
-        """
-        generated_code = RenderConfigVisitor().render_node(self)
-        # TODO: Instead, maybe configure rendering without colors?
-        generated_code = self._erase_annotation_colors(generated_code)
-        compiled_code: CodeType = compile(generated_code, "<string>", "exec")
-        exec_namespace = {}
-        exec(compiled_code, exec_namespace)
-        return exec_namespace[self.node_name]
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    @staticmethod
-    def _erase_annotation_colors(
-        generated_code: str,
-    ):
-        for term_color in [
-            TermColor.config_comment,
-            TermColor.config_missing,
-            TermColor.config_unused,
-            TermColor.reset_style,
-        ]:
-            generated_code = generated_code.replace(term_color.value, "")
-        return generated_code
-
-
-class AbstractConfLeapRootNode(AbstractRootNode):
-    """
-    Base implementation for all `ConfLeap.*`.
-    """
-
-    def __init__(
-        self,
-        conf_leap: ConfLeap,
-        child_builder: ConfigBuilderVisitor,
-        **kwargs,
-    ):
-        super().__init__(
-            node_name=conf_leap.name,
-            is_present=True,
-            child_builder=child_builder,
-            **kwargs,
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-class AbstractConfLeapNodeBuilder(ConfigBuilderVisitor):
-
-    @staticmethod
-    def _create_used_dict_field(
-        dict_node: AbstractDictNode,
-        field_name: str,
-        node_class: type,
-        conf_leap: ConfLeap,
-        **kwargs,
-    ) -> AbstractDictNode:
-        field_name = field_name
-        kwargs.pop("is_present", None)
-        kwargs.pop("orig_data", None)
-        kwargs.pop("node_name", None)
-        kwargs.pop("node_indent", None)
-        field_node: AbstractDictNode = node_class(
-            node_name=field_name,
-            node_indent=dict_node.node_indent + AbstractConfigNode.indent_size,
-            is_present=(field_name in dict_node.orig_data),
-            orig_data=dict_node.orig_data.get(field_name, None),
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        dict_node.child_nodes[field_name] = field_node
-        return field_node
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def _create_common_fields(
-        self,
-        dict_node: AbstractDictNode,
-        conf_leap: ConfLeap,
-    ):
-        # Common overridable `global` and `local` fields: FT_23_37_64_44.conf_dst.md
-
-        self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=ConfField.field_required_python_version.value,
-            node_class=Node_field_required_python_version,
-            conf_leap=conf_leap,
-        )
-
-        self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=ConfField.field_python_selector_file_rel_path.value,
-            node_class=Node_field_python_selector_file_rel_path,
-            conf_leap=conf_leap,
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=ConfField.field_local_venv_dir_rel_path.value,
-            node_class=Node_field_local_venv_dir_rel_path,
-            conf_leap=conf_leap,
-        )
-
-        self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=ConfField.field_local_log_dir_rel_path.value,
-            node_class=Node_field_local_log_dir_rel_path,
-            conf_leap=conf_leap,
-        )
-
-        self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=ConfField.field_local_tmp_dir_rel_path.value,
-            node_class=Node_field_local_tmp_dir_rel_path,
-            conf_leap=conf_leap,
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=ConfField.field_local_cache_dir_rel_path.value,
-            node_class=Node_field_local_cache_dir_rel_path,
-            conf_leap=conf_leap,
-        )
-
-        self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=ConfField.field_venv_driver.value,
-            node_class=Node_field_venv_driver,
-            conf_leap=conf_leap,
-        )
-
-        self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=ConfField.field_project_descriptors.value,
-            node_class=Node_field_project_descriptors,
-            conf_leap=conf_leap,
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    @staticmethod
-    def _create_unused_dict_fields(
-        dict_node: AbstractDictNode,
-    ):
-        for field_name, field_value in dict_node.orig_data.items():
-
-            if field_name not in dict_node.child_nodes:
-                dict_node.child_nodes[field_name] = UnusedConfigBuilderVisitor().build_config_node(
-                    node_name=field_name,
-                    node_indent=dict_node.node_indent + AbstractConfigNode.indent_size,
-                    orig_data=field_value,
-                )
-
-
-########################################################################################################################
-# `ConfLeap.leap_input` node types.
-# See: FT_19_44_42_19.effective_config.md
-
-
-# noinspection PyPep8Naming
-class Builder_RootNode_input(AbstractConfLeapNodeBuilder):
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def visit_dict(
-        self,
-        dict_node: AbstractDictNode,
-        **kwargs,
-    ) -> None:
-
-        conf_leap = ConfLeap.leap_input
-
-        field_node: AbstractConfigNode
-
-        field_node = self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_proto_code_file_abs_path_inited.name,
-            node_class=AbstractValueNode,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        field_node.note_text = (
-            f"Value `{EnvState.state_proto_code_file_abs_path_inited.name}` is an absolute path to `{ConfConstGeneral.name_proto_code}`.\n"
-            f"It allows resolving all other relative paths (via `{PathName.path_ref_root.value}` - see field `{ConfField.field_ref_root_dir_rel_path.value}`).\n"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        )
-
-        field_node = self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_primer_conf_file_abs_path_inited.name,
-            node_class=AbstractValueNode,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        # TODO: Link to `ConfLeap.leap_derived` fields.
-        field_node.note_text = (
-            f"Value `{EnvState.state_primer_conf_file_abs_path_inited.name}` is an absolute path to `{ConfLeap.leap_primer}` config file.\n"
-            f"The config file is selected from the list of possible candidates (whichever is found first, replacing extension to `.{ConfConstInput.conf_file_ext}`):\n"
-            f"*   basename of the entry script,\n"
-            f"*   basename of the `{ConfConstGeneral.name_proto_code}` file,\n"
-            f"*   default `{ConfConstInput.default_file_basename_conf_primer}`.\n"
-            f"Note that the selected config file basename is subsequently re-used for others:\n"
-            f"*   see `{EnvState.state_global_conf_file_abs_path_inited.name}` for `{ConfLeap.leap_client.name}`,\n"
-            f"*   see `{EnvState.state_local_conf_file_abs_path_inited.name}` for `{ConfLeap.leap_env.name}`.\n"
-            #
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        self._create_unused_dict_fields(dict_node)
-
-
-# noinspection PyPep8Naming
-class RootNode_input(AbstractConfLeapRootNode):
-    """
-    Root node for `ConfLeap.leap_input`.
-    """
-
-    def __init__(
-        self,
-        **kwargs,
-    ):
-        super().__init__(
-            conf_leap=ConfLeap.leap_input,
-            child_builder=Builder_RootNode_input(),
-            **kwargs,
-        )
-        self.note_text = (
-            f"The `{ConfLeap.leap_input.name}` data is taken from the `{ConfConstGeneral.name_proto_code}` process input (not configured in files):\n"
-            f"*   CLI args, environment variables, current directory, ...\n"
-            f"*   combination of the above with applied defaults.\n"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        )
-
-
-########################################################################################################################
-# `ConfLeap.leap_primer` node types.
-# See: FT_19_44_42_19.effective_config.md
-
-
-# noinspection PyPep8Naming
-class Builder_RootNode_primer(AbstractConfLeapNodeBuilder):
-
-    def __init__(
-        self,
-        state_primer_conf_file_abs_path_inited: str,
-    ):
-        self.state_primer_conf_file_abs_path_inited: str = state_primer_conf_file_abs_path_inited
-
-    def visit_dict(
-        self,
-        dict_node: AbstractDictNode,
-        **kwargs,
-    ) -> None:
-        self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=ConfField.field_ref_root_dir_rel_path.value,
-            node_class=Node_field_ref_root_dir_rel_path,
-            state_primer_conf_file_abs_path_inited=self.state_primer_conf_file_abs_path_inited,
-            conf_leap=ConfLeap.leap_primer,
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=ConfField.field_global_conf_dir_rel_path.value,
-            node_class=Node_field_global_conf_dir_rel_path,
-            state_primer_conf_file_abs_path_inited=self.state_primer_conf_file_abs_path_inited,
-            conf_leap=ConfLeap.leap_primer,
-        )
-
-        self._create_unused_dict_fields(dict_node)
-
-
-# noinspection PyPep8Naming
-class RootNode_primer(AbstractConfLeapRootNode):
-    """
-    Root node for `ConfLeap.leap_primer`.
-    """
-
-    def __init__(
-        self,
-        state_primer_conf_file_abs_path_inited: str,
-        **kwargs,
-    ):
-        super().__init__(
-            conf_leap=ConfLeap.leap_primer,
-            child_builder=Builder_RootNode_primer(
-                state_primer_conf_file_abs_path_inited=state_primer_conf_file_abs_path_inited,
-            ),
-            **kwargs,
-        )
-        self.state_primer_conf_file_abs_path_inited: str = state_primer_conf_file_abs_path_inited
-        self.note_text = f"The `{ConfLeap.leap_primer.name}` data is loaded from the [{self.state_primer_conf_file_abs_path_inited}] file."
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-# noinspection PyPep8Naming
-class Node_field_ref_root_dir_rel_path(AbstractValueNode[str]):
-
-    def __init__(
-        self,
-        state_primer_conf_file_abs_path_inited: str,
-        **kwargs,
-    ):
-        super().__init__(
-            **kwargs,
-        )
-        self.state_primer_conf_file_abs_path_inited: str = state_primer_conf_file_abs_path_inited
-        self.note_text = (
-            f"Field `{ConfField.field_ref_root_dir_rel_path.value}` points to the dir called `{PathName.path_ref_root.value}`.\n"
-            f"The path is relative to the `{ConfConstGeneral.name_proto_code}` file [{self.state_primer_conf_file_abs_path_inited}].\n"
-            f"Normally, the `{PathName.path_ref_root.value}` dir is the client repo root, but it can be anything.\n"
-            f"See `{EnvState.state_ref_root_dir_abs_path_inited.name}` in `{ConfLeap.leap_derived.name}` -\n"
-            f"the derived abs path is the base path for all the configured relative paths (except for this field itself, obviously).\n"
-            #
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-# noinspection PyPep8Naming
-class Node_field_global_conf_dir_rel_path(AbstractValueNode[str]):
-
-    def __init__(
-        self,
-        state_primer_conf_file_abs_path_inited: str,
-        **kwargs,
-    ):
-        super().__init__(
-            **kwargs,
-        )
-        self.state_primer_conf_file_abs_path_inited: str = state_primer_conf_file_abs_path_inited
-        self.note_text = (
-            f"Field `{ConfField.field_global_conf_dir_rel_path.value}` points to the global config dir (as opposed to local config dir `{ConfField.field_local_conf_symlink_rel_path.value}`).\n"
-            f"{ConfConstGeneral.relative_path_field_note}\n"
-            f"See `{EnvState.state_global_conf_dir_abs_path_inited.name}` in `{ConfLeap.leap_derived.name}` -\n"
-            f"normally, the resolved global config dir contains all other global client config files.\n"
-            #
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-########################################################################################################################
-# `ConfLeap.leap_client` node types.
-# See: FT_19_44_42_19.effective_config.md
-
-
-# noinspection PyPep8Naming
-class Builder_RootNode_client(AbstractConfLeapNodeBuilder):
-
-    def visit_dict(
-        self,
-        dict_node: AbstractDictNode,
-        **kwargs,
-    ) -> None:
-
-        conf_leap = ConfLeap.leap_client
-
-        self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=ConfField.field_local_conf_symlink_rel_path.value,
-            node_class=Node_field_local_conf_symlink_rel_path,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=ConfField.field_default_env_dir_rel_path.value,
-            node_class=Node_field_default_env_dir_rel_path,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-
-        self._create_common_fields(
-            dict_node=dict_node,
-            conf_leap=conf_leap,
-        )
-
-        self._create_unused_dict_fields(dict_node)
-
-
-# noinspection PyPep8Naming
-class RootNode_client(AbstractConfLeapRootNode):
-    """
-    Root node for `ConfLeap.leap_client`.
-    """
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def __init__(
-        self,
-        state_global_conf_file_abs_path_inited: str,
-        **kwargs,
-    ):
-        super().__init__(
-            conf_leap=ConfLeap.leap_client,
-            child_builder=Builder_RootNode_client(),
-            **kwargs,
-        )
-        self.state_global_conf_file_abs_path_inited: str = state_global_conf_file_abs_path_inited
-        self.note_text = f"The `{ConfLeap.leap_client.name}` data is loaded from the [{self.state_global_conf_file_abs_path_inited}] file."
-
-
-# noinspection PyPep8Naming
-class Node_field_local_conf_symlink_rel_path(AbstractValueNode[str]):
-
-    def __init__(
-        self,
-        **kwargs,
-    ):
-        super().__init__(
-            **kwargs,
-        )
-        self.note_text = (
-            f"Field `{ConfField.field_local_conf_symlink_rel_path.value}` points to local config dir (as opposed to the global config dir `{ConfField.field_global_conf_dir_rel_path.value}`).\n"
-            f"{ConfConstGeneral.relative_path_field_note}\n"
-            f"The basename of this path is a symlink set to the actual dir with environment-specific config.\n"
-            f"If the symlink does not exist yet, its target is set from:\n"
-            f"*   either field `{ConfField.field_default_env_dir_rel_path.value}`,\n"
-            f"*   or arg `{SyntaxArg.arg_env}` which can also be used to re-set the symlink target to a new path.\n"
-            f"See `{EnvState.state_global_conf_dir_abs_path_inited.name}` in `{ConfLeap.leap_derived.name}` -\n"
-            f"normally, the resolved local config dir contains all local environment-specific config files.\n"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        )
-
-
-# noinspection PyPep8Naming
-class Node_field_default_env_dir_rel_path(AbstractValueNode[str]):
-
-    def __init__(
-        self,
-        **kwargs,
-    ):
-        super().__init__(
-            **kwargs,
-        )
-        self.note_text = (
-            f"Field `{ConfField.field_default_env_dir_rel_path.value}` is the default path where `{ConfField.field_local_conf_symlink_rel_path.value}` symlink can point to.\n"
-            f"{ConfConstGeneral.relative_path_field_note}\n"
-            f"The path is ignored when the `{ConfField.field_local_conf_symlink_rel_path.value}` symlink already exists.\n"
-            f"Arg `{SyntaxArg.arg_env}` overrides this `{ConfField.field_default_env_dir_rel_path.value}` field.\n"
-            #
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-# noinspection PyPep8Naming
-class Node_field_required_python_version(AbstractValueNode[str]):
-
-    def __init__(
-        self,
-        conf_leap: ConfLeap,
-        **kwargs,
-    ):
-        super().__init__(
-            **kwargs,
-        )
-        if conf_leap == ConfLeap.leap_client:
-            self.note_text = (
-                f"Field `{ConfField.field_required_python_version.value}` selects `python` version.\n"
-                f'The value specifies the version of `python` interpreter which is used to create `venv` (e.g. "{ConfConstEnv.latest_known_python_version}").\n'
-                f"{ConfConstGeneral.common_field_global_note}\n"
-                #
-            )
-        elif conf_leap == ConfLeap.leap_env:
-            self.note_text = f"{ConfConstGeneral.common_field_local_note}\n"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-# noinspection PyPep8Naming
-class Node_field_python_selector_file_rel_path(AbstractValueNode[str]):
-
-    def __init__(
-        self,
-        conf_leap: ConfLeap,
-        **kwargs,
-    ):
-        super().__init__(
-            **kwargs,
-        )
-        if conf_leap == ConfLeap.leap_client:
-            self.note_text = (
-                f"Field `{ConfField.field_python_selector_file_rel_path.value}` specifies rel path to `python` selector.\n"
-                f"The selector is a standalone script written in `python` which must implement `{SelectorFunc.select_python_file_abs_path.value}` function.\n"
-                f"{ConfConstGeneral.common_field_global_note}\n"
-                #
-            )
-        elif conf_leap == ConfLeap.leap_env:
-            self.note_text = f"{ConfConstGeneral.common_field_local_note}\n"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-# noinspection PyPep8Naming
-class Node_field_local_venv_dir_rel_path(AbstractValueNode[str]):
-
-    def __init__(
-        self,
-        conf_leap: ConfLeap,
-        **kwargs,
-    ):
-        super().__init__(
-            **kwargs,
-        )
-        if conf_leap == ConfLeap.leap_client:
-            self.note_text = (
-                f"Field `{ConfField.field_local_venv_dir_rel_path.value}` points to the dir where `venv` (`python` virtual environment) is created.\n"
-                f"{ConfConstGeneral.relative_path_field_note}\n"
-                f"{ConfConstGeneral.common_field_global_note}\n"
-                #
-            )
-        elif conf_leap == ConfLeap.leap_env:
-            self.note_text = f"{ConfConstGeneral.common_field_local_note}\n"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-# noinspection PyPep8Naming
-class Node_field_local_log_dir_rel_path(AbstractValueNode[str]):
-
-    def __init__(
-        self,
-        conf_leap: ConfLeap,
-        **kwargs,
-    ):
-        super().__init__(
-            **kwargs,
-        )
-        if conf_leap == ConfLeap.leap_client:
-            self.note_text = (
-                f"Field `{ConfField.field_local_log_dir_rel_path.value}` points to the dir with log files created for each script execution.\n"
-                f"{ConfConstGeneral.relative_path_field_note}\n"
-                f"{ConfConstGeneral.common_field_global_note}\n"
-                #
-            )
-        elif conf_leap == ConfLeap.leap_env:
-            self.note_text = f"{ConfConstGeneral.common_field_local_note}\n"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-# noinspection PyPep8Naming
-class Node_field_local_tmp_dir_rel_path(AbstractValueNode[str]):
-
-    def __init__(
-        self,
-        conf_leap: ConfLeap,
-        **kwargs,
-    ):
-        super().__init__(
-            **kwargs,
-        )
-        if conf_leap == ConfLeap.leap_client:
-            self.note_text = (
-                f"Field `{ConfField.field_local_tmp_dir_rel_path.value}` points to the dir with temporary files created for some commands.\n"
-                f"{ConfConstGeneral.relative_path_field_note}\n"
-                f"{ConfConstGeneral.common_field_global_note}\n"
-                #
-            )
-        elif conf_leap == ConfLeap.leap_env:
-            self.note_text = f"{ConfConstGeneral.common_field_local_note}\n"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-# noinspection PyPep8Naming
-class Node_field_local_cache_dir_rel_path(AbstractValueNode[str]):
-
-    def __init__(
-        self,
-        conf_leap: ConfLeap,
-        **kwargs,
-    ):
-        super().__init__(
-            **kwargs,
-        )
-        if conf_leap == ConfLeap.leap_client:
-            self.note_text = (
-                f"Field `{ConfField.field_local_cache_dir_rel_path.value}` points to the dir with cached files created for some commands.\n"
-                f"{ConfConstGeneral.relative_path_field_note}\n"
-                f"{ConfConstGeneral.common_field_global_note}\n"
-                #
-            )
-        elif conf_leap == ConfLeap.leap_env:
-            self.note_text = f"{ConfConstGeneral.common_field_local_note}\n"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-# noinspection PyPep8Naming
-class Node_field_venv_driver(AbstractValueNode[str]):
-
-    def __init__(
-        self,
-        conf_leap: ConfLeap,
-        **kwargs,
-    ):
-        super().__init__(
-            **kwargs,
-        )
-        if conf_leap == ConfLeap.leap_client:
-            self.note_text = (
-                f"Field `{ConfField.field_venv_driver.value}` selects a tool to manage packages:\n"
-                f'*   specify "{VenvDriverType.venv_pip.name}" to use native `pip`,\n'
-                f'*   specify "{VenvDriverType.venv_uv.name}" to use fast `uv`.\n'
-                f"{ConfConstGeneral.common_field_global_note}\n"
-                #
-            )
-        elif conf_leap == ConfLeap.leap_env:
-            self.note_text = f"{ConfConstGeneral.common_field_local_note}\n"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-# noinspection PyPep8Naming
-class Builder_Node_field_project_descriptors(AbstractConfLeapNodeBuilder):
-
-    def build_dict_node(
-        self,
-        node_name: str | None,
-        node_indent: int,
-        orig_data: dict,
-        conf_leap: ConfLeap,
-        **kwargs,
-    ) -> "AbstractDictNode":
-        return Node_project_descriptor(
-            node_indent=node_indent,
-            orig_data=orig_data,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-
-
-# noinspection PyPep8Naming
-class Node_field_project_descriptors(AbstractListNode):
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def __init__(
-        self,
-        conf_leap: ConfLeap,
-        **kwargs,
-    ):
-        super().__init__(
-            conf_leap=conf_leap,
-            child_builder=Builder_Node_field_project_descriptors(),
-            **kwargs,
-        )
-        if conf_leap == ConfLeap.leap_client:
-            self.note_text = (
-                f"Field `{ConfField.field_project_descriptors.value}` lists `python` projects and their installation details.\n"
-                f"{ConfConstGeneral.common_field_global_note}\n"
-                # See: UC_78_58_06_54.no_stray_packages.md:
-                f"Note that the `{ConfConstGeneral.name_protoprimer_package}` does not manage package dependencies itself.\n"
-                f"Instead, the `{ConfConstGeneral.name_protoprimer_package}` relies on `{ConfConstClient.default_pyproject_toml_basename}` file per `python` project to specify these dependencies.\n"
-                f"See `{EnvState.state_project_descriptors_inited.name}` in `{ConfLeap.leap_derived.name}`.\n"
-            )
-        elif conf_leap == ConfLeap.leap_env:
-            self.note_text = f"{ConfConstGeneral.common_field_local_note}\n"
-        elif conf_leap == ConfLeap.leap_derived:
-            self.note_text = f"{ConfConstGeneral.func_note_derived_based_on_common(ConfField.field_project_descriptors.value)}\n"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-# noinspection PyPep8Naming
-class Builder_Node_project_descriptor(AbstractConfLeapNodeBuilder):
-
-    def __init__(
-        self,
-        conf_leap: ConfLeap,
-        **kwargs,
-    ):
-        super().__init__(
-            **kwargs,
-        )
-        self.conf_leap: ConfLeap = conf_leap
-
-    def visit_dict(
-        self,
-        dict_node: AbstractDictNode,
-        **kwargs,
-    ) -> None:
-        self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=ConfField.field_build_root_dir_rel_path.value,
-            node_class=Node_field_build_root_dir_rel_path,
-            conf_leap=self.conf_leap,
-            **kwargs,
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=ConfField.field_install_extras.value,
-            node_class=Node_field_install_extras,
-            conf_leap=self.conf_leap,
-            **kwargs,
-        )
-
-        self._create_unused_dict_fields(dict_node)
-
-
-# noinspection PyPep8Naming
-class Node_project_descriptor(AbstractDictNode):
-
-    def __init__(
-        self,
-        conf_leap: ConfLeap,
-        **kwargs,
-    ):
-        kwargs.pop("node_name", None)
-        kwargs.pop("is_present", None)
-        super().__init__(
-            node_name=None,
-            is_present=True,
-            child_builder=Builder_Node_project_descriptor(
-                conf_leap=conf_leap,
-            ),
-            **kwargs,
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-# noinspection PyPep8Naming
-class Node_field_build_root_dir_rel_path(AbstractValueNode[str]):
-
-    def __init__(
-        self,
-        conf_leap: ConfLeap,
-        **kwargs,
-    ):
-        super().__init__(
-            **kwargs,
-        )
-        if conf_leap in [
-            ConfLeap.leap_client,
-            ConfLeap.leap_derived,
-        ]:
-            self.note_text = (
-                f"This is similar to specifying the dir of `{ConfConstClient.default_pyproject_toml_basename}` for `pip`:\n"
-                f"pip install path/to/project\n"
-                f"{ConfConstGeneral.relative_path_field_note}\n"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-            )
-
-
-# noinspection PyPep8Naming
-class Node_field_install_extras(AbstractListNode):
-
-    def __init__(
-        self,
-        conf_leap: ConfLeap,
-        **kwargs,
-    ):
-        super().__init__(
-            child_builder=ConfigBuilderVisitor(),
-            **kwargs,
-        )
-        if conf_leap in [
-            ConfLeap.leap_client,
-            ConfLeap.leap_derived,
-        ]:
-            self.note_text = (
-                f"This is similar to specifying a list of `extra_item`-s per `path/to/project` for `pip`:\n"
-                f"pip install path/to/project[extra_item_1,extra_item_2,...]\n"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-            )
-
-
-########################################################################################################################
-# `ConfLeap.leap_env` node types.
-# See: FT_19_44_42_19.effective_config.md
-
-
-# noinspection PyPep8Naming
-class Builder_RootNode_env(AbstractConfLeapNodeBuilder):
-
-    def visit_dict(
-        self,
-        dict_node: AbstractDictNode,
-        **kwargs,
-    ) -> None:
-
-        self._create_common_fields(
-            dict_node=dict_node,
-            conf_leap=ConfLeap.leap_env,
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        self._create_unused_dict_fields(dict_node)
-
-
-# noinspection PyPep8Naming
-class RootNode_env(AbstractConfLeapRootNode):
-    """
-    Root node for `ConfLeap.leap_env`.
-    """
-
-    def __init__(
-        self,
-        state_local_conf_file_abs_path_inited: str,
-        **kwargs,
-    ):
-        super().__init__(
-            conf_leap=ConfLeap.leap_env,
-            child_builder=Builder_RootNode_env(),
-            **kwargs,
-        )
-        self.state_local_conf_file_abs_path_inited: str = state_local_conf_file_abs_path_inited
-        self.note_text = f"The `{ConfLeap.leap_env.name}` data is loaded from the [{self.state_local_conf_file_abs_path_inited}] file."
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-
-########################################################################################################################
-# `ConfLeap.leap_derived` node types.
-# See: FT_19_44_42_19.effective_config.md
-
-
-# noinspection PyPep8Naming
-class Builder_RootNode_derived(AbstractConfLeapNodeBuilder):
-
-    def visit_dict(
-        self,
-        dict_node: AbstractDictNode,
-        **kwargs,
-    ) -> None:
-
-        conf_leap = ConfLeap.leap_derived
-
-        field_node: AbstractConfigNode
-
-        # ===
-        # `ConfLeap.leap_input`
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        field_node = self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_proto_code_file_abs_path_inited.name,
-            node_class=AbstractValueNode,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        field_node.note_text = f"{ConfConstGeneral.func_note_derived_based_on_conf_leap_field(EnvState.state_proto_code_file_abs_path_inited.name, ConfLeap.leap_input)}\n"
-
-        field_node = self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_primer_conf_file_abs_path_inited.name,
-            node_class=AbstractValueNode,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        field_node.note_text = f"{ConfConstGeneral.func_note_derived_based_on_conf_leap_field(EnvState.state_primer_conf_file_abs_path_inited.name, ConfLeap.leap_input)}\n"
-
-        # ===
-        # `ConfLeap.leap_primer`
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        field_node = self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_ref_root_dir_abs_path_inited.name,
-            node_class=AbstractValueNode,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        field_node.note_text = f"{ConfConstGeneral.func_note_derived_based_on_conf_leap_field(ConfField.field_ref_root_dir_rel_path.value, ConfLeap.leap_primer)}\n"
-
-        field_node = self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_global_conf_dir_abs_path_inited.name,
-            node_class=AbstractValueNode,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        field_node.note_text = f"{ConfConstGeneral.func_note_derived_based_on_conf_leap_field(ConfField.field_global_conf_dir_rel_path.value, ConfLeap.leap_primer)}\n"
-
-        field_node = self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_global_conf_file_abs_path_inited.name,
-            node_class=AbstractValueNode,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        field_node.note_text = (
-            # TODO: It is not derived from just this:
-            #       *   dirname is from `field_global_conf_dir_rel_path`
-            #       *   basename is from `state_primer_conf_file_abs_path_inited`
-            f"{ConfConstGeneral.func_note_derived_based_on_conf_leap_field(EnvState.state_primer_conf_file_abs_path_inited.name, ConfLeap.leap_input)}\n"
-        )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        # ===
-        # `ConfLeap.leap_client`
-
-        field_node = self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_selected_env_dir_rel_path_inited.name,
-            node_class=AbstractValueNode,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        field_node.note_text = (
-            # TODO: Either default or --env arg:
-            f"{ConfConstGeneral.func_note_derived_based_on_conf_leap_field(ConfField.field_default_env_dir_rel_path.value, ConfLeap.leap_client)}\n"
-        )
-
-        field_node = self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_local_conf_symlink_abs_path_inited.name,
-            node_class=AbstractValueNode,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        field_node.note_text = f"{ConfConstGeneral.func_note_derived_based_on_conf_leap_field(ConfField.field_local_conf_symlink_rel_path.value, ConfLeap.leap_client)}\n"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        field_node = self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_local_conf_file_abs_path_inited.name,
-            node_class=AbstractValueNode,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        field_node.note_text = (
-            # TODO: It is not derived from just this:
-            #       *   dirname is from `field_local_conf_symlink_rel_path`
-            #       *   basename is from `state_primer_conf_file_abs_path_inited`
-            f"{ConfConstGeneral.func_note_derived_based_on_conf_leap_field(EnvState.state_primer_conf_file_abs_path_inited.name, ConfLeap.leap_input)}\n"
-        )
-
-        # ===
-        # `ConfLeap.leap_env`
-        # nothing specific
-
-        # ===
-        # `ConfLeap.leap_derived`
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        field_node = self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_required_python_version_inited.name,
-            node_class=AbstractValueNode,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        field_node.note_text = f"{ConfConstGeneral.func_note_derived_based_on_common(ConfField.field_required_python_version.value)}\n"
-
-        field_node = self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_python_selector_file_abs_path_inited.name,
-            node_class=AbstractValueNode,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        field_node.note_text = f"{ConfConstGeneral.func_note_derived_based_on_common(ConfField.field_python_selector_file_rel_path.value)}\n"
-
-        field_node = self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_local_venv_dir_abs_path_inited.name,
-            node_class=AbstractValueNode,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        field_node.note_text = f"{ConfConstGeneral.func_note_derived_based_on_common(ConfField.field_local_venv_dir_rel_path.value)}\n"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        field_node = self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_local_log_dir_abs_path_inited.name,
-            node_class=AbstractValueNode,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        field_node.note_text = f"{ConfConstGeneral.func_note_derived_based_on_common(ConfField.field_local_log_dir_rel_path.value)}\n"
-
-        field_node = self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_local_tmp_dir_abs_path_inited.name,
-            node_class=AbstractValueNode,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        field_node.note_text = f"{ConfConstGeneral.func_note_derived_based_on_common(ConfField.field_local_tmp_dir_rel_path.value)}\n"
-
-        field_node = self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_local_cache_dir_abs_path_inited.name,
-            node_class=AbstractValueNode,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        field_node.note_text = f"{ConfConstGeneral.func_note_derived_based_on_common(ConfField.field_local_cache_dir_rel_path.value)}\n"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        field_node = self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_venv_driver_inited.name,
-            node_class=AbstractValueNode,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-        field_node.note_text = f"{ConfConstGeneral.func_note_derived_based_on_common(ConfField.field_venv_driver.value)}\n"
-
-        self._create_used_dict_field(
-            dict_node=dict_node,
-            field_name=EnvState.state_project_descriptors_inited.name,
-            node_class=Node_field_project_descriptors,
-            conf_leap=conf_leap,
-            **kwargs,
-        )
-
-        self._create_unused_dict_fields(dict_node)
-
-
-# noinspection PyPep8Naming
-class RootNode_derived(AbstractConfLeapRootNode):
-    """
-    Root node for `ConfLeap.leap_derived`.
-    """
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-    def __init__(
-        self,
-        **kwargs,
-    ):
-        super().__init__(
-            conf_leap=ConfLeap.leap_derived,
-            child_builder=Builder_RootNode_derived(),
-            **kwargs,
-        )
-        self.note_text = (
-            f"The `{ConfLeap.leap_derived.name}` data is derived from other data - it is computed by:\n"
-            f"*   applying defaults to missing field values\n"
-            f"*   combining with other field values\n"
-            f"Effectively, this is what ultimately used by the `{ConfConstGeneral.name_protoprimer_package}`.\n"
-            #
-        )
-
-
-########################################################################################################################
-
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
 class RunStrategy:
     """
     See related:
@@ -3208,7 +1642,7 @@ class RunStrategy:
           Currently, `RunStrategy` is degenerated into single implementation `ExitCodeReporter`.
           Is it even needed (unless make it useful beyond that)?
     """
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
     def execute_strategy(
         self,
         state_node: StateNode,
@@ -3220,7 +1654,7 @@ class ExitCodeReporter(RunStrategy):
     """
     This strategy requires the state to return an `int` value (as exit code).
     """
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
     def __init__(
         self,
         env_ctx: EnvContext,
@@ -3234,7 +1668,7 @@ class ExitCodeReporter(RunStrategy):
     ) -> None:
         """
         This is a trivial implementation.
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
         No special DAG traversal because nodes traverse their own dependencies.
         But it may not reach all nodes because
         dependencies will be conditionally evaluated by the implementation of those nodes.
@@ -3244,7 +1678,7 @@ class ExitCodeReporter(RunStrategy):
         if type(exit_code) is not int:
             raise AssertionError("`exit_code` must be an `int`")
         sys.exit(exit_code)
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
 
 ########################################################################################################################
 
@@ -3262,10 +1696,10 @@ class StateNode(Generic[ValueType]):
     ):
         self.env_ctx: EnvContext = env_ctx
         self.state_name: str = state_name
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
         # Ensure no duplicates:
         assert len(parent_states) == len(set(parent_states))
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
         self.parent_states: list[str] = parent_states
 
         assert type(state_name) is str
@@ -3997,34 +2431,32 @@ class Bootstrapper_state_primer_conf_file_data_loaded(AbstractCachingStateNode[d
 
         if can_print_effective_config(self):
 
-            state_input_stderr_log_level_eval_finalized: int = self.eval_parent_state(EnvState.state_input_stderr_log_level_eval_finalized.name)
-            is_quiet: bool = state_input_stderr_log_level_eval_finalized > logging.INFO
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-            if is_quiet:
-                # Print this note early
-                logging.warning(f"Showing plain config data - use option `-v` (`--verbose`) to annotate it with descriptions.")
-
             # Print `ConfLeap.leap_input` data together:
             # ===
             # `ConfLeap.leap_input`:
-            conf_input = RootNode_input(
-                node_indent=0,
-                orig_data={
-                    EnvState.state_proto_code_file_abs_path_inited.name: state_proto_code_file_abs_path_inited,
-                    EnvState.state_primer_conf_file_abs_path_inited.name: state_primer_conf_file_abs_path_inited,
-                },
+            print(
+                json.dumps(
+                    {
+                        ConfLeap.leap_input.name: {
+                            EnvState.state_proto_code_file_abs_path_inited.name: state_proto_code_file_abs_path_inited,
+                            EnvState.state_primer_conf_file_abs_path_inited.name: state_primer_conf_file_abs_path_inited,
+                        }
+                    },
+                    indent=4,
+                )
             )
-            print(RenderConfigVisitor(is_quiet=is_quiet).render_node(conf_input))
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
             # ===
             # `ConfLeap.leap_primer`:
-            conf_primer = RootNode_primer(
-                node_indent=0,
-                orig_data=file_data,
-                state_primer_conf_file_abs_path_inited=state_primer_conf_file_abs_path_inited,
+            print(
+                json.dumps(
+                    {
+                        ConfLeap.leap_primer.name: file_data,
+                    },
+                    indent=4,
+                )
             )
-            print(RenderConfigVisitor(is_quiet=is_quiet).render_node(conf_primer))
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
         return file_data
 
 
@@ -4039,14 +2471,14 @@ class Bootstrapper_state_ref_root_dir_abs_path_inited(AbstractCachingStateNode[s
         ]
     )
     _state_name = staticmethod(lambda: EnvState.state_ref_root_dir_abs_path_inited.name)
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
     def _eval_state_once(
         self,
     ) -> ValueType:
         state_proto_code_file_abs_path_inited = self.eval_parent_state(EnvState.state_proto_code_file_abs_path_inited.name)
 
         proto_code_dir_abs_path: str = os.path.dirname(state_proto_code_file_abs_path_inited)
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
         state_primer_conf_file_data_loaded: dict = self.eval_parent_state(EnvState.state_primer_conf_file_data_loaded.name)
 
         field_client_dir_rel_path: str | None = state_primer_conf_file_data_loaded.get(
@@ -4066,9 +2498,9 @@ class Bootstrapper_state_ref_root_dir_abs_path_inited(AbstractCachingStateNode[s
                 proto_code_dir_abs_path,
                 field_client_dir_rel_path,
             )
-
-        state_ref_root_dir_abs_path_inited = os.path.normpath(state_ref_root_dir_abs_path_inited)
 ########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+        state_ref_root_dir_abs_path_inited = os.path.normpath(state_ref_root_dir_abs_path_inited)
+
         assert os.path.isabs(state_ref_root_dir_abs_path_inited)
         return state_ref_root_dir_abs_path_inited
 
@@ -4088,9 +2520,9 @@ class Bootstrapper_state_global_conf_dir_abs_path_inited(AbstractCachingStateNod
     def _eval_state_once(
         self,
     ) -> ValueType:
-
-        state_ref_root_dir_abs_path_inited: str = self.eval_parent_state(EnvState.state_ref_root_dir_abs_path_inited.name)
 ########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+        state_ref_root_dir_abs_path_inited: str = self.eval_parent_state(EnvState.state_ref_root_dir_abs_path_inited.name)
+
         state_primer_conf_file_data_loaded: dict = self.eval_parent_state(EnvState.state_primer_conf_file_data_loaded.name)
 
         field_client_config_dir_rel_path: str | None = state_primer_conf_file_data_loaded.get(
@@ -4109,9 +2541,9 @@ class Bootstrapper_state_global_conf_dir_abs_path_inited(AbstractCachingStateNod
                 state_ref_root_dir_abs_path_inited,
                 field_client_config_dir_rel_path,
             )
-
-        return state_global_conf_dir_abs_path_inited
 ########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+        return state_global_conf_dir_abs_path_inited
+
 
 # noinspection PyPep8Naming
 @trivial_factory
@@ -4131,9 +2563,9 @@ class Bootstrapper_state_global_conf_file_abs_path_inited(AbstractCachingStateNo
 
         state_primer_conf_file_abs_path_inited: str = self.eval_parent_state(EnvState.state_primer_conf_file_abs_path_inited.name)
         conf_file_base_name = os.path.basename(state_primer_conf_file_abs_path_inited)
-
-        state_global_conf_dir_abs_path_inited: str = self.eval_parent_state(EnvState.state_global_conf_dir_abs_path_inited.name)
 ########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+        state_global_conf_dir_abs_path_inited: str = self.eval_parent_state(EnvState.state_global_conf_dir_abs_path_inited.name)
+
         state_global_conf_file_abs_path_inited: str = os.path.join(
             state_global_conf_dir_abs_path_inited,
             conf_file_base_name,
@@ -4173,16 +2605,15 @@ class Bootstrapper_state_client_conf_file_data_loaded(AbstractCachingStateNode[d
             file_data = {}
 
         if can_print_effective_config(self):
-            state_input_stderr_log_level_eval_finalized: int = self.eval_parent_state(EnvState.state_input_stderr_log_level_eval_finalized.name)
-            is_quiet: bool = state_input_stderr_log_level_eval_finalized > logging.INFO
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-            conf_client = RootNode_client(
-                node_indent=0,
-                orig_data=file_data,
-                state_global_conf_file_abs_path_inited=state_global_conf_file_abs_path_inited,
+            print(
+                json.dumps(
+                    {
+                        ConfLeap.leap_client.name: file_data,
+                    },
+                    indent=4,
+                )
             )
-            print(RenderConfigVisitor(is_quiet=is_quiet).render_node(conf_client))
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
         return file_data
 
 
@@ -4198,7 +2629,7 @@ class Bootstrapper_state_selected_env_dir_rel_path_inited(AbstractCachingStateNo
         ]
     )
     _state_name = staticmethod(lambda: EnvState.state_selected_env_dir_rel_path_inited.name)
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
     def _eval_state_once(
         self,
     ) -> ValueType:
@@ -4206,7 +2637,7 @@ class Bootstrapper_state_selected_env_dir_rel_path_inited(AbstractCachingStateNo
         client_local_env_dir_any_path: str | None = self._select_client_local_env_dir_any_path()
         if client_local_env_dir_any_path is None:
             return None
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
         client_local_env_dir_abs_path: str = self._select_client_local_env_dir_abs_path(client_local_env_dir_any_path)
 
         if not os.path.isdir(client_local_env_dir_abs_path):
@@ -4225,9 +2656,9 @@ class Bootstrapper_state_selected_env_dir_rel_path_inited(AbstractCachingStateNo
                 state_ref_root_dir_abs_path_inited,
             )
         )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-        assert ".." not in pathlib.Path(state_selected_env_dir_rel_path_inited).parts
 
+        assert ".." not in pathlib.Path(state_selected_env_dir_rel_path_inited).parts
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
         return state_selected_env_dir_rel_path_inited
 
     def _select_client_local_env_dir_any_path(
@@ -4435,16 +2866,15 @@ class Bootstrapper_state_env_conf_file_data_loaded(AbstractCachingStateNode[dict
             file_data = {}
 
         if can_print_effective_config(self):
-            state_input_stderr_log_level_eval_finalized: int = self.eval_parent_state(EnvState.state_input_stderr_log_level_eval_finalized.name)
-            is_quiet: bool = state_input_stderr_log_level_eval_finalized > logging.INFO
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
-            conf_env = RootNode_env(
-                node_indent=0,
-                orig_data=file_data,
-                state_local_conf_file_abs_path_inited=state_local_conf_file_abs_path_inited,
+            print(
+                json.dumps(
+                    {
+                        ConfLeap.leap_env.name: file_data,
+                    },
+                    indent=4,
+                )
             )
-            print(RenderConfigVisitor(is_quiet=is_quiet).render_node(conf_env))
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
         return file_data
 
 
@@ -4460,7 +2890,7 @@ class Bootstrapper_required_python_version_inited(AbstractOverriddenFieldCaching
         ]
     )
     _state_name = staticmethod(lambda: EnvState.state_required_python_version_inited.name)
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
     def _eval_state_once(
         self,
     ) -> ValueType:
@@ -4468,7 +2898,7 @@ class Bootstrapper_required_python_version_inited(AbstractOverriddenFieldCaching
             ConfField.field_required_python_version.value,
             None,
         )
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
         state_ref_root_dir_abs_path_inited: str = self.eval_parent_state(EnvState.state_ref_root_dir_abs_path_inited.name)
 
         if state_required_python_version_inited is None:
@@ -4484,14 +2914,14 @@ class Bootstrapper_required_python_version_inited(AbstractOverriddenFieldCaching
         # normalize:
         python_version: tuple[int, int, int] = parse_python_version(state_required_python_version_inited)
         state_required_python_version_inited = f"{python_version[0]}.{python_version[1]}.{python_version[2]}"
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
         return state_required_python_version_inited
 
 
 # noinspection PyPep8Naming
 @trivial_factory
 class Bootstrapper_state_python_selector_file_abs_path_inited(AbstractOverriddenFieldCachingStateNode[str]):
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
     _parent_states = staticmethod(
         lambda: [
             EnvState.state_ref_root_dir_abs_path_inited.name,
@@ -4509,7 +2939,7 @@ class Bootstrapper_state_python_selector_file_abs_path_inited(AbstractOverridden
             ConfField.field_python_selector_file_rel_path.value,
             None,
         )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
         state_python_selector_file_abs_path_inited: str | None
         if python_selector_file_rel_path is not None:
             state_ref_root_dir_abs_path_inited: str = self.eval_parent_state(EnvState.state_ref_root_dir_abs_path_inited.name)
@@ -4519,7 +2949,7 @@ class Bootstrapper_state_python_selector_file_abs_path_inited(AbstractOverridden
             )
         else:
             state_python_selector_file_abs_path_inited = None
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
         return state_python_selector_file_abs_path_inited
 
 
@@ -4536,11 +2966,11 @@ class Bootstrapper_state_selected_python_file_abs_path_inited(AbstractCachingSta
         ]
     )
     _state_name = staticmethod(lambda: EnvState.state_selected_python_file_abs_path_inited.name)
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
     def _eval_state_once(
         self,
     ) -> ValueType:
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
         state_python_selector_file_abs_path_inited: str | None = self.eval_parent_state(EnvState.state_python_selector_file_abs_path_inited.name)
 
         state_required_python_version_inited: str = self.eval_parent_state(EnvState.state_required_python_version_inited.name)
@@ -4558,7 +2988,7 @@ class Bootstrapper_state_selected_python_file_abs_path_inited(AbstractCachingSta
 # noinspection PyPep8Naming
 @trivial_factory
 class Bootstrapper_state_local_venv_dir_abs_path_inited(AbstractOverriddenFieldCachingStateNode[str]):
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
     _parent_states = staticmethod(
         lambda: [
             EnvState.state_ref_root_dir_abs_path_inited.name,
@@ -4567,7 +2997,7 @@ class Bootstrapper_state_local_venv_dir_abs_path_inited(AbstractOverriddenFieldC
         ]
     )
     _state_name = staticmethod(lambda: EnvState.state_local_venv_dir_abs_path_inited.name)
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
     def _eval_state_once(
         self,
     ) -> ValueType:
@@ -4584,11 +3014,11 @@ class Bootstrapper_state_local_venv_dir_abs_path_inited(AbstractOverriddenFieldC
                 state_ref_root_dir_abs_path_inited,
                 state_local_venv_dir_abs_path_inited,
             )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
         assert os.path.isabs(state_local_venv_dir_abs_path_inited)
         return state_local_venv_dir_abs_path_inited
 
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
 # noinspection PyPep8Naming
 @trivial_factory
 class Bootstrapper_state_local_log_dir_abs_path_inited(AbstractOverriddenFieldCachingStateNode[str]):
@@ -4605,12 +3035,12 @@ class Bootstrapper_state_local_log_dir_abs_path_inited(AbstractOverriddenFieldCa
     def _eval_state_once(
         self,
     ) -> ValueType:
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
         field_local_log_dir_rel_path: str = self._get_overridden_value_or_default(
             ConfField.field_local_log_dir_rel_path.value,
             ConfConstEnv.default_dir_rel_path_log,
         )
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
         state_ref_root_dir_abs_path_inited: str = self.eval_parent_state(EnvState.state_ref_root_dir_abs_path_inited.name)
 
         state_local_log_dir_abs_path_inited = os.path.join(
@@ -4626,7 +3056,7 @@ class Bootstrapper_state_local_log_dir_abs_path_inited(AbstractOverriddenFieldCa
 # noinspection PyPep8Naming
 @trivial_factory
 class Bootstrapper_state_local_tmp_dir_abs_path_inited(AbstractOverriddenFieldCachingStateNode[str]):
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
     _parent_states = staticmethod(
         lambda: [
             EnvState.state_ref_root_dir_abs_path_inited.name,
@@ -4635,7 +3065,7 @@ class Bootstrapper_state_local_tmp_dir_abs_path_inited(AbstractOverriddenFieldCa
         ]
     )
     _state_name = staticmethod(lambda: EnvState.state_local_tmp_dir_abs_path_inited.name)
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
     def _eval_state_once(
         self,
     ) -> ValueType:
@@ -4652,11 +3082,11 @@ class Bootstrapper_state_local_tmp_dir_abs_path_inited(AbstractOverriddenFieldCa
             field_local_tmp_dir_rel_path,
         )
         state_local_tmp_dir_abs_path_inited = os.path.normpath(state_local_tmp_dir_abs_path_inited)
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
         assert os.path.isabs(state_local_tmp_dir_abs_path_inited)
         return state_local_tmp_dir_abs_path_inited
 
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
 # noinspection PyPep8Naming
 @trivial_factory
 class Bootstrapper_state_local_cache_dir_abs_path_inited(AbstractOverriddenFieldCachingStateNode[str]):
@@ -4673,12 +3103,12 @@ class Bootstrapper_state_local_cache_dir_abs_path_inited(AbstractOverriddenField
     def _eval_state_once(
         self,
     ) -> ValueType:
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
         field_local_cache_dir_rel_path: str = self._get_overridden_value_or_default(
             ConfField.field_local_cache_dir_rel_path.value,
             ConfConstEnv.default_dir_rel_path_cache,
         )
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
         state_ref_root_dir_abs_path_inited: str = self.eval_parent_state(EnvState.state_ref_root_dir_abs_path_inited.name)
 
         state_local_cache_dir_abs_path_inited = os.path.join(
@@ -4694,7 +3124,7 @@ class Bootstrapper_state_local_cache_dir_abs_path_inited(AbstractOverriddenField
 # noinspection PyPep8Naming
 @trivial_factory
 class Bootstrapper_state_venv_driver_inited(AbstractOverriddenFieldCachingStateNode[VenvDriverType]):
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
     _parent_states = staticmethod(
         lambda: [
             EnvState.state_client_conf_file_data_loaded.name,
@@ -4703,7 +3133,7 @@ class Bootstrapper_state_venv_driver_inited(AbstractOverriddenFieldCachingStateN
         ]
     )
     _state_name = staticmethod(lambda: EnvState.state_venv_driver_inited.name)
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
     def _eval_state_once(
         self,
     ) -> ValueType:
@@ -4719,7 +3149,7 @@ class Bootstrapper_state_venv_driver_inited(AbstractOverriddenFieldCachingStateN
             default_venv_driver = VenvDriverType.venv_pip.name
         else:
             default_venv_driver = ConfConstEnv.default_venv_driver
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
         state_venv_driver_inited: VenvDriverType
         venv_driver_name: str | None = os.environ.get(EnvVar.var_PROTOPRIMER_VENV_DRIVER.value, None)
         if venv_driver_name is None:
@@ -4731,7 +3161,7 @@ class Bootstrapper_state_venv_driver_inited(AbstractOverriddenFieldCachingStateN
             ]
         else:
             state_venv_driver_inited = VenvDriverType[venv_driver_name]
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
         if (
             selected_version < uv_min_version
             and state_venv_driver_inited == VenvDriverType.venv_uv
@@ -4741,7 +3171,7 @@ class Bootstrapper_state_venv_driver_inited(AbstractOverriddenFieldCachingStateN
             state_venv_driver_inited = VenvDriverType.venv_pip
 
         return state_venv_driver_inited
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
 
 # noinspection PyPep8Naming
 @trivial_factory
@@ -4754,7 +3184,7 @@ class Bootstrapper_state_project_descriptors_inited(AbstractOverriddenFieldCachi
         ]
     )
     _state_name = staticmethod(lambda: EnvState.state_project_descriptors_inited.name)
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
     def _eval_state_once(
         self,
     ) -> ValueType:
@@ -4763,7 +3193,7 @@ class Bootstrapper_state_project_descriptors_inited(AbstractOverriddenFieldCachi
             ConfField.field_project_descriptors.value,
             ConfConstEnv.default_project_descriptors,
         )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
         return project_descriptors
 
 
@@ -4778,7 +3208,7 @@ class Bootstrapper_state_install_specs_inited(AbstractOverriddenFieldCachingStat
         ]
     )
     _state_name = staticmethod(lambda: EnvState.state_install_specs_inited.name)
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
     def _eval_state_once(
         self,
     ) -> ValueType:
@@ -4787,7 +3217,7 @@ class Bootstrapper_state_install_specs_inited(AbstractOverriddenFieldCachingStat
             ConfField.field_install_specs.value,
             ConfConstEnv.default_install_specs,
         )
-########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
+
         return install_specs
 
 
@@ -4799,7 +3229,7 @@ class Bootstrapper_state_derived_conf_data_loaded(AbstractCachingStateNode[dict]
     """
 
     _state_name = staticmethod(lambda: EnvState.state_derived_conf_data_loaded.name)
-
+########### !!!!! GENERATED CONTENT - ANY CHANGES WILL BE LOST !!!!! ###########
     def __init__(
         self,
         env_ctx: EnvContext,
@@ -4865,14 +3295,14 @@ class Bootstrapper_state_derived_conf_data_loaded(AbstractCachingStateNode[dict]
                 config_data_derived[derived_data_env_state] = evaluated_value
 
         if can_print_effective_config(self):
-            state_input_stderr_log_level_eval_finalized: int = self.eval_parent_state(EnvState.state_input_stderr_log_level_eval_finalized.name)
-            is_quiet: bool = state_input_stderr_log_level_eval_finalized > logging.INFO
-
-            conf_derived = RootNode_derived(
-                node_indent=0,
-                orig_data=config_data_derived,
+            print(
+                json.dumps(
+                    {
+                        ConfLeap.leap_derived.name: config_data_derived,
+                    },
+                    indent=4,
+                )
             )
-            print(RenderConfigVisitor(is_quiet=is_quiet).render_node(conf_derived))
 
         return config_data_derived
 
@@ -5951,6 +4381,7 @@ class StateGraph:
     ) -> StateNode:
         if state_name not in self.state_nodes:
             self.state_nodes[state_name] = self.state_factories[state_name].create_state_node(env_ctx)
+            logger.debug(f"instantiated `state_node` class [{self.state_nodes[state_name].__class__}] for `state_name` [{state_name}]")
         return self.state_nodes[state_name]
 
     def eval_state(
