+++ src/neoprimer/main/neoprimer/cmd_venv_shell.py
@@ -12,6 +12,8 @@ logger = logging.getLogger()
 
 
 def custom_main():
+    # TODO: TODO_28_48_19_20.api_to_traverse_config_when_primed.md:
+    #       convert from `boot_env` to `start_app` when it can access the config.
     proto_main(customize_env_context)
