+++ src/neoprimer/test/test_neoprimer/test_conf_renderer/test_RootNode_client.py
@@ -2,11 +2,13 @@ from __future__ import annotations
 
 from local_test.integrated_helper import test_python_version
 from local_test.name_assertion import assert_test_module_name_embeds_str
+from neoprimer.conf_renderer import (
+    RenderConfigVisitor,
+    RootNode_client,
+)
 from protoprimer.primer_kernel import (
     ConfConstEnv,
     ConfField,
-    RenderConfigVisitor,
-    RootNode_client,
     TermColor,
 )
