+++ src/neoprimer/test/test_neoprimer/test_conf_renderer/test_RootNode_derived.py
@@ -1,11 +1,13 @@
 from __future__ import annotations
 
 from local_test.name_assertion import assert_test_module_name_embeds_str
+from neoprimer.conf_renderer import (
+    RenderConfigVisitor,
+    RootNode_derived,
+)
 from protoprimer.primer_kernel import (
     ConfField,
     EnvState,
-    RenderConfigVisitor,
-    RootNode_derived,
     TermColor,
 )
