+++ src/neoprimer/test/test_neoprimer/test_conf_renderer/test_RootNode_input.py
@@ -1,12 +1,14 @@
 from __future__ import annotations
 
 from local_test.name_assertion import assert_test_module_name_embeds_str
+from neoprimer.conf_renderer import (
+    RenderConfigVisitor,
+    RootNode_input,
+)
 from protoprimer.primer_kernel import (
     ConfConstGeneral,
     ConfConstInput,
     EnvState,
-    RenderConfigVisitor,
-    RootNode_input,
     TermColor,
 )
