+++ src/neoprimer/test/test_neoprimer/test_fast_slim_max_mocked/test_state_bootstrapper/test_Bootstrapper_state_activated_venv_shell_started.py
@@ -32,7 +32,7 @@ class ThisTestClass(BasePyfakefsTestClass):
 
     # noinspection PyMethodMayBeStatic
     def test_relationship(self):
-        assert_test_module_name_embeds_str(Bootstrapper_state_activated_venv_shell_started._state_name())
+        assert_test_module_name_embeds_str(Bootstrapper_state_activated_venv_shell_started.__name__)
 
     @patch("sys.argv", [""])
     @patch(f"{primer_kernel.__name__}.{Bootstrapper_state_args_parsed.__name__}.create_state_node")
